+++ crates/board/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.10.0-git
+## 0.10.0
 
 ### Major
 
@@ -217,4 +217,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 10 -->
+<!-- Increment to skip CHANGELOG.md test: 0 -->